remove timestamp causing double JSON serialization
For some reason having this extra timestamp here caused a double serialization of the `@timestamp` field in the output. We were getting two because beats originally wanted you to set your own timestamp, but later changed the library to include it automatically as it was a required field for every event, when we upgraded the backing beats library this remained.